More progress
Added some integration tests to show that everything is hooked up to
receieve channel messages and server messages from the local dev server.
This also showed I was missing a possible type of message: the user
list. Added vague support for it, will need to come back later when I
have example text of a longer user list to be more complete.

Also started to clean up the demo view controller to use the new code
instead of the old experimental stuff.